+++ webapp/src/main/java/rocks/metaldetector/web/api/response/NotificationConfigResponse.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import rocks.metaldetector.web.api.EmailConfig;
-import rocks.metaldetector.web.api.TelegramConfig;
 
 @Data
 @NoArgsConstructor
